pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "regressionlab"
version = "1.1.2"
description = "Scientific curve fitting application with GUI"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Alejandro Mata Ali", email = "alejandro.mata.ali@gmail.com"}
]
keywords = ["curve-fitting", "scientific", "data-analysis"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "numpy>=2.0,<3.0",
    "matplotlib>=3.10,<4.0",
    "scipy>=1.17,<2.0",
    "pandas>=2.3,<3.0",
    "openpyxl>=3.1,<4.0",
    "Pillow>=10.0,<11.0",
    "python-dotenv>=1.0,<2.0",
    "colorama>=0.4,<1.0",
    "PyYAML>=6.0,<7.0",
    "streamlit>=1.31,<2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0,<8.0",
    "pytest-cov>=4.0,<5.0",
    "black>=23.0,<24.0",
    "ruff>=0.1,<1.0",
    "mypy>=1.0,<2.0",
    "pre-commit>=3.0,<4.0",
]
docs = [
    "sphinx>=9.0.0",
    "sphinx-rtd-theme>=3.0.0",
    "myst-parser>=5.0.0",
    "sphinx-autodoc-typehints>=3.0.0",
    "linkify-it-py>=2.0.0",
]

[project.scripts]
regressionlab = "regressionlab.main_program:main"
regressionlab-streamlit = "regressionlab.streamlit_runner:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "scripts/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
